fixed terminal light
